[nodejs-review] review-4 fixed. refactor project structure - move app.js in src - change server.js to bin/www executable - npm start for prod, npm run dev for develop environment
