+++ projects/koajs-proj-conf/bin/www
@@ -1,6 +1,8 @@
+#!/usr/bin/env node
+
 const http = require('http');
 
-const app = require('./app');
+const app = require('../src/app');
 
 
 const server = http.createServer(app.callback());
