+++ projects/koajs-proj-conf/docker-compose.yml
@@ -11,7 +11,7 @@ services:
       - "8000:8000"
     volumes:
       - "./package.json:/opt/service/package.json"
-      - "./app.js:/opt/service/app.js"
+      - "./bin:/opt/service/bin"
       - "./src:/opt/service/src"
       - ".eslintrc:/opt/service/.eslintrc"
     environment:
