+++ projects/koajs-proj-conf/src/api.test.js
@@ -1,5 +1,5 @@
 const request = require('supertest');
-const app = require('../app');
+const app = require('./app');
 const db = require('./db');
 const { API_VER } = require('./config');
